MA-743 Added `BluetoothScaleSetupStore` to manage the multi-step flow for Bluetooth scale setup.

+++ iOS/meApp/Core/Shared/Utilities/ConversionTools.swift
@@ -78,9 +78,9 @@ final class ConversionTools {
     }
     
     /// Converts Bluetooth scale kg value to stored value (tenths of lbs)
-    static func convertBluetoothToStored(_ btKg: Double) -> Int {
+    static func convertBluetoothToStored(_ btKg: Double) -> Double {
         // Bluetooth scales: .2 lbs resolution, special formula
-        return Int(rounded(btKg * 1.1023 * 2 * 10, toPlaces: 0) / 10)
+        return Double(rounded(btKg * 1.1023 * 2 * 10, toPlaces: 0) / 10)
     }
     
     /// Appsync display to stored (kg to lbs, more precision)
